refactor(vscode-ail-chat): reorganize process/ into platform subfolders
Move process killers into src/process/{posix,win}/killer.ts and
factory into src/process/factory.ts. Tests follow the same layout.
Establishes the subfolder pattern for future platform-specific
process management code.
