+++ vscode-ail-chat/src/process/killer.ts
