+++ vscode-ail-chat/test/process/win/killer.test.ts
@@ -4,10 +4,8 @@ import * as path from 'path';
 
 describe('WindowsProcessKiller', () => {
   it('calls taskkill with /F /T /PID <pid>', () => {
-    // Verify the argument construction — pid is always numeric so no injection possible.
-    // We inspect the source rather than spawning a real process to keep tests platform-safe.
     const source = fs.readFileSync(
-      path.join(__dirname, '../../src/process/windows-process-killer.ts'),
+      path.join(__dirname, '../../../src/process/win/killer.ts'),
       'utf-8'
     );
     expect(source).toContain("'taskkill'");
